Move implementation to splines module
